mpfrlint: update about various tools.
Run all the tools present in dist-hook of Makefile.am, in the
same order (note that currently, tools/ck-inits-clears is also
run by mpfrlint); doc/check-typography is run earlier, to group
it with other doc tests.

For tools/ck-mparam, no longer check for gcc as it is no longer
needed as of commit 6636bf7424bb7571b9be186b34842832a52e1621
(2018-04-03).