[doc/mpfr.texi] Use @node with only one argument, as recommended.
The Texinfo manual says:

  However, if your Texinfo document is hierarchically organized,
  as virtually all are, we recommend leaving off all the pointers
  and letting ‘texi2any’ determine them.

This is the case for the MPFR manual. According to tests, the only
change is that for each first subsection, the "Previous" pointer,
which was the same as the "Up" pointer, is no longer present. This
is more consistent with the last subsection, which did not have a
"Next" node. When using 'p' and 'n' in info, this also avoids one
mistakenly leaving the current section level.